Add ugly placeholder launch screens
Unfortunately, neither of these is very effective: the coloured
background is displayed for a short time, then the screen clears
to white, then we eventually get our desired content. I think
this means that Polymer is taking a significant time to load;
so if we want placeholder content it will probably have to be HTML.

It seems we also get the portrait configuration (blue) in either
orientation, never the landscape configuration (red).